Modified narrow synonym indexing in MP core.

+++ data-model-solr/src/main/java/org/mousephenotype/cda/solr/service/dto/MpDTO.java
@@ -324,7 +324,7 @@ public class MpDTO {
 	private List<String> mgiAccessionId;
 
 	@Field(PHENO_CALLS)
-	private Integer phenoCalls;
+	private Long phenoCalls;
 	
 	@Field(MARKER_SYMBOL)
 	private List<String> markerSymbol;
@@ -1196,12 +1196,12 @@ public class MpDTO {
 		this.mgiAccessionId = mgiAccessionId;
 	}
 
-	public Integer getPhenoCalls() {
+	public Long getPhenoCalls() {
 
 		return phenoCalls;
 	}
 	
-	public void setPhenoCalls(Integer mpCalls) {
+	public void setPhenoCalls(Long mpCalls) {
 		
 		this.phenoCalls = mpCalls;
 	}
